Branch 64bit merget to the trunk
git-svn-id: http://merlin.fit.vutbr.cz/svn/STK/trunk@73 1d61b05a-ac01-0410-9a19-bfe7c875843c

+++ src/SLRatio.cc
@@ -437,7 +437,7 @@ int main(int argc, char *argv[]) {
                             derivOrder, derivWinLengths, &header,
                             cmn_path, cvn_path, cvg_file, &rhfbuff);
 
-    if (hset.mInputVectorSize != header.mSampleSize / sizeof(float)) {
+    if (hset.mInputVectorSize != static_cast<int>(header.mSampleSize / sizeof(float))) {
       Error("Vector size [%d] in '%s' is incompatible with HMM set [%d]",
             header.mSampleSize/sizeof(float), file_name->mpPhysical, hset.mInputVectorSize);
     }
